list-objects-filter: add filter profile plugin support
Filter Profile Plugins are external libraries that implement filters
using the interface defined in list-objects-filter-profile.c to create
custom filtering behaviours.

This is not a stable binary interface and plugins will need to be
recompiled alongside Git. Since the normal use cases of filtering is in
remote fetch operations this would only required on the server-side.

All profile filters are implemented using the new internal 'profile'
filter, which handles loading and unloading the library, context, and
provides a lightweight wrapper around the internal filter interface.

When requested via --filter=profile:<name>[=<arg>], the <name> profile
filter is resolved from the configuration variable
'uploadfilter.profile:<name>.plugin' to a library path. If not absolute,
the library path is resolved from GIT_EXEC_DIR. The plugin is loaded via
dlopen() and the three functions implementing the plugin are loaded via
dlsym().

1. git_filter_profile_init() is called to construct any necessary
   context and parse and validate any <arg> provided to the profile
   filter.

2. git_filter_profile_object() is called from
   list_objects_filter__filter_object() to implement the main filter
   function. To simplify plugin implementation any oidset manipulation
   is not passed to the profile filter but is performed in the caller,
   triggered with a flag.

3. git_filter_profile_free() is called to clean up any context.

To implement a profile filter plugin, #include
list-objects-filter-profile.h and implement the three functions above.
Since the interface exposed is fairly minimal, the plugins can be
implemented in non-C environments.

Signed-off-by: Robert Coup <robert.coup@koordinates.com>

RC/move config parse