Makefile: add support for building filter profile plugins
Define FILTER_PROFILES as a space-separated list of absolute or relative
paths to build as filter profile plugins. Plugins are expected to have a
Makefile with all/install/clean tasks. Several variables are passed to
the profile makefiles.

Also enables linking with -rdynamic or equivalent to guarantee that
dlopen()'d plugins can resolve symbols from the main application.

Signed-off-by: Robert Coup <robert.coup@koordinates.com>