filter-profiles: add example filter plugin implementations and readme
The rand profile filter returns all trees and a %age of the blobs
(specified by the filter argument), randomly selected. The necessary
build and linking behaviour is included in Makefile.

The rand-cpp profile filter has the same behaviour as rand but is
implemented in C++17, and the associated build/link info in Makefile.

Signed-off-by: Robert Coup <robert.coup@koordinates.com>